requirements.txt
# Core Dependencies
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
numpy>=1.26.0
pandas>=2.0.0

# AI/ML Libraries
ultralytics>=8.0.0  # YOLOv8
torch>=2.0.0
torchvision>=0.15.0

# Object Tracking
filterpy>=1.4.5
scikit-learn>=1.3.0
scipy>=1.11.0

# UI Framework
streamlit>=1.28.0

# Visualization
matplotlib>=3.8.0
plotly>=5.17.0
seaborn>=0.13.0

# Data Processing
Pillow>=10.0.0
PyYAML>=6.0.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
